feat.delete dir  user_group
